Analytics Drawing Basis
This thing is running fast - Just under 30fps for N > 100,000

That can't be right, it's way too fast! Investigate

+++ OOT_2/OOT_2/ParticleSystem.h
@@ -13,7 +13,7 @@ public:
 	ParticleSystem();
 
 	//Intialises with a specific number of particles
-	void InitWith(unsigned int _n);
+	void InitWith(size_t _n);
 
 	void Clear();
